requirements.txt
Flask==0.12.2
Jinja2==2.9.6
google-cloud-storage==1.4.0
Pillow==4.2.1
requests==2.18.4
gunicorn==19.7.1
graphene>=2.0.dev
firebase-admin==2.3.0
google-cloud-vision==0.27.0
webcolors==1.3
opencv-python==3.3.0.10
matplotlib==2.1.0
yapf==0.22.0
